backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx         # if you use it elsewhere
requests      # <— add this line
